feat: New game state
