+++ src/games/sequence/module/players/monte_carlo.py
@@ -1,5 +1,5 @@
 from .player import BasePlayer
-from .utils.alphazero import encode
+from .utils.game import encode
 from .utils.mc import monte_carlo, rollout_maker, selector_maker
