--- src/games/sequence/module/players/utils/__init__.py
@@ -1,3 +1,0 @@
-from .alphazero import *
-from .mc import *
-from .game import *
