+++ src/games/sequence/module/players/utils/alphazero.py
@@ -1,80 +1,8 @@
-from typing import Tuple
-from ...utils import BoardViewer
-from ...defaults import ALL_CARDS_MAPPING, CORNERS, JACK
 from ...sequence import Sequence
-from .types import Action, GameData, Position, RolloutMaker, Selector, State, Encoder, Any, Dict, List, Card
-from .game import calc_colab
+from .types import Action, RolloutMaker, Selector, State, Encoder, Any, Dict, List, Tuple
+from .game import calc_colab, encode_valids
 from math import sqrt
 import numpy as np
-
-
-def table_bit(i: int, j: int) -> int:
-    return i * 10 + j
-
-
-def encode_board(board: BoardViewer, color: Any = -1) -> int:
-    masks = [0]
-    colors = {color: 0}
-    for (i, j), item in board:
-        if item:
-            if not item.color in colors:
-                colors[item.color] = len(colors)
-                masks.append(0)
-            masks[colors[item.color]] |= 1 << table_bit(i, j)           
-    return masks
-
-
-def encode_cards(cards: List[Card], foo) -> int:
-    mask = 0
-    data = {c:0 for c in cards}
-    for c in cards:
-        mask |= 1 << table_bit(*ALL_CARDS_MAPPING[c][data[c]])
-        data[c] += 1
-    return mask
-
-
-def adjust_shifting(pos: Position) -> int:
-    return len([1 for x in CORNERS if x < pos])
-
-
-def encode_valids(valids: List[Action]) -> int:
-    if valids[0] is None:
-        return 1 << 198
-    mask = 0
-    discards = 0
-    for (_, num), pos in valids:
-        if pos is None:
-            mask |= 1 << (192 + discards)
-            discards += 1
-        else:
-            cur_bit = 1 << (table_bit(*pos) - adjust_shifting(pos))
-            if num is JACK: mask |= cur_bit << 96
-            else : mask |= cur_bit
-    return mask
-
-
-def encode(
-    player: GameData,
-    discard_pile: List[Card],
-) -> State :       
-    boards = encode_board(player.board, player.color)
-    cards = encode_cards(list(player.cards), "player")
-    pile = encode_cards(discard_pile, "pile")
-    offset = 0
-    state = 0
-    for mask in [*boards, cards, pile]:
-        state += (mask << offset)
-        offset += 104
-    return state
-
-    
-def state_to_list(
-    state: State,
-    size: int,
-) -> List[int]:
-    binary_rep = bin(state)[2:]
-    binary_rep = '0' * max(0, size - len(binary_rep)) + binary_rep
-    return [int(x) for x in binary_rep[-1 : -(size + 1) : -1]]
 
 
 def rollout_maker(
